requirements.txt
opencv-python
deepface
pafy
youtube-dl
python-vlc
